requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-admin-honeypot==1.1.0
django-redis==5.3.0
djangorestframework==3.14.0
Faker==18.10.1
Pillow==9.5.0
